net: hns3: Fix the double unmap issue while reliquishing the buffer
For the non-XDP scenario,  buffer unmap happens when buffer is freed or replaced.
Currently this does not checks for the condition if buffer was already unampped
(as driver cannot reuse the buffer) because either it was being passed to the
stack, XDP_TX'ed or it was XDP_REDIRECT'ed.

Buffer unmap can happen outside of the existing hns3_buffer_unmap() function for
both non-XDP and XDP scenarios. Hence, we need to check if the buffer was
already unmapped earlier.

Tested-by: Hao Chen <chenhao418@huawei.com>
Signed-off-by: Hao Chen <chenhao418@huawei.com>
Signed-off-by: Jian Shen <shenjian15@huawei.com>
Signed-off-by: Salil Mehta <salil.mehta@huawei.com>