net: hns3: Fix the check for Max limit of GRO packet in non-XDP path
Current non-XDP path only checks for the FE Bit to detect the EOP. This can lead
to infinite loop and system can hang. This should be limited by the clean count
for this NAPI cycle and maximum number of BDs in a GRO packet supported by
hardware. If the later check fails then packet should be dropped. This is a
pathological check in software for buggy hardware.

Signed-off-by: Hao Chen <chenhao418@huawei.com>
Signed-off-by: Jian Shen <shenjian15@huawei.com>
Signed-off-by: Salil Mehta <salil.mehta@huawei.com>